+++ backend/src/main/java/kernel360/techpick/feature/link/service/dto/LinkResponse.java
@@ -1,4 +1,4 @@
-package kernel360.techpick.feature.link.model.dto;
+package kernel360.techpick.feature.link.service.dto;
 
 public record LinkResponse(Long id, String url, String title, String description) {
 }
