+++ backend/src/main/java/kernel360/techpick/feature/tag/model/TagMapper.java
@@ -5,8 +5,8 @@ import org.springframework.stereotype.Component;
 import kernel360.techpick.feature.user.exception.ApiUserException;
 import kernel360.techpick.core.model.tag.Tag;
 import kernel360.techpick.core.model.user.User;
-import kernel360.techpick.feature.tag.model.dto.TagCreateRequest;
-import kernel360.techpick.feature.tag.model.dto.TagResponse;
+import kernel360.techpick.feature.tag.service.dto.TagCreateRequest;
+import kernel360.techpick.feature.tag.service.dto.TagResponse;
 import kernel360.techpick.feature.user.UserRepository;
 import lombok.RequiredArgsConstructor;
