+++ backend/src/main/java/kernel360/techpick/feature/tag/service/dto/TagUpdateRequest.java
@@ -1,4 +1,4 @@
-package kernel360.techpick.feature.tag.model.dto;
+package kernel360.techpick.feature.tag.service.dto;
 
 public record TagUpdateRequest(
 	Long id,
